+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Resource product type: made `attribute` elements optional
  - Resource product type: Validate/protect `required` element on Product type attribute
  - Resource type: Validate/protect `required` element on Type attribute
+ - Resource type: Avoid `changeAttributeOrder` update action when new attribute gets added
  - Resource product type: Avoid `changeAttributeOrder` update action when new attribute gets added
  - Resource product type: Added support for Nested types